requirements.txt
streamlit>=1.32.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
fpdf2>=2.7.0
plotly>=5.18.0
pandas>=2.0.0
